feat(output): add ResultFormatter for JSON/YAML output
Introduce ResultFormatter to handle formatting and output of test
results in JSON or YAML based on configuration. This centralizes output
logic, supports writing to file or stdout, and improves error handling.

refactor: remove hostname verification and simplify SSLTest

Remove custom hostname verification logic and unused imports from
SSLTest. Delegate certificate validation and output formatting to
dedicated classes for cleaner, modular code structure.
