+++ app/.settings/org.eclipse.buildship.core.prefs
@@ -1,13 +1,13 @@
-arguments=--init-script /Users/windy/.cache/nvim/jdtls/app/config/org.eclipse.osgi/155/0/.cp/gradle/init/init.gradle
+arguments=
 auto.sync=false
 build.scans.enabled=false
-connection.gradle.distribution=GRADLE_DISTRIBUTION(LOCAL_INSTALLATION(/Users/windy/.sdkman/candidates/gradle/current))
-connection.project.dir=
+connection.gradle.distribution=GRADLE_DISTRIBUTION(WRAPPER)
+connection.project.dir=..
 eclipse.preferences.version=1
 gradle.user.home=
-java.home=/Users/windy/.sdkman/candidates/java/21.0.5-graal
+java.home=
 jvm.arguments=
 offline.mode=false
-override.workspace.settings=true
-show.console.view=true
-show.executions.view=true
+override.workspace.settings=false
+show.console.view=false
+show.executions.view=false
