+++ app/bin/main/logback.xml
@@ -6,7 +6,7 @@
         </encoder>
     </appender>
 
-    <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
+    <!-- <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
         <file>logs/ssltest.log</file>
         <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
             <fileNamePattern>logs/ssltest.%d{yyyy-MM-dd}.log</fileNamePattern>
@@ -15,10 +15,10 @@
         <encoder>
             <pattern>%d{yyyy-MM-dd HH:mm:ss} %-5level - %msg%n</pattern>
         </encoder>
-    </appender>
+    </appender> -->
 
     <root level="INFO">
         <appender-ref ref="CONSOLE" />
-        <appender-ref ref="FILE" />
+        <!-- <appender-ref ref="FILE" /> -->
     </root>
 </configuration>
