+++ app/src/main/java/org/example/output/ResultFormatter.java
@@ -1,0 +1,55 @@
+package org.example.output;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import org.example.config.SSLTestConfig;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.FileWriter;
+import java.io.PrintWriter;
+import java.util.Map;
+
+public class ResultFormatter {
+    private static final Logger logger = LoggerFactory.getLogger(ResultFormatter.class);
+    private final ObjectMapper jsonMapper = new ObjectMapper();
+    private final ObjectMapper yamlMapper = new ObjectMapper(new YAMLFactory());
+    private final SSLTestConfig config;
+
+    public ResultFormatter(SSLTestConfig config) {
+        this.config = config;
+    }
+
+    public void formatAndOutput(Map<String, Object> result) {
+        try {
+            String output;
+            switch (config.getFormat()) {
+                case JSON:
+                    output = jsonMapper.writerWithDefaultPrettyPrinter().writeValueAsString(result);
+                    break;
+                case YAML:
+                    output = yamlMapper.writeValueAsString(result);
+                    break;
+                default:
+                    return; // TEXT format is handled by direct logging
+            }
+
+            if (config.getOutputFile() != null) {
+                try (PrintWriter writer = new PrintWriter(new FileWriter(config.getOutputFile()))) {
+                    writer.println(output);
+                }
+            } else {
+                System.out.println(output);
+            }
+        } catch (Exception e) {
+            logger.error("输出结果时发生错误: {}", e.getMessage());
+        }
+    }
+
+    public void logError(String message, Throwable cause, int exitCode) {
+        logger.error("❌ {}", message);
+        if (cause != null && config.isVerbose()) {
+            logger.error("详细错误信息:", cause);
+        }
+    }
+}
